Notebooks/requirements.txt
pyyaml>=5.3.1
numpy>=1.18.4
opencv-python>=4.2.0.32
scikit-image>=0.17.2
scipy>=1.4.1
pillow>=7.1.2
matplotlib>=3.1.3
tensorflow>=2.3.0
efficientnet>=1.1.0
regex>=2020.6.8
scikit-learn>=0.22.1
seaborn>=0.10.0
pandas>=0.25.3
six>=1.15.0
tqdm>=4.46.0
psycopg2>=2.8.4
timezonefinder>=4.4.0
pytz>=2020.1
geopy>=2.0.0
xhtml2pdf>=0.2.4
jinja2>=2.11.2
pypdf2>=1.26.0
reportlab>=3.5.42
imgaug>=0.2.9
networkx>=2.4
requests>=2.23.0
